+++ tests/test_tool_annotations.py
@@ -48,7 +48,7 @@ DELETE_CLASS_TOOLS = {
 WRITE_TOOLS = {
     "accept_candidate", "ask", "auto_review_trigger", "broadcast",
     "candidate_review_run", "claim_pickup", "close_thread", "config_reload",
-    "convene_panel", "core_set", "curator_review", "dialectic_claim",
+    "convene_panel", "core_set", "curator_review", "db_compact", "dialectic_claim",
     "dialectic_evidence", "dialectic_mine_run", "dialectic_observation_resolve",
     "dialectic_supersede", "dialectic_validate_run", "distill", "evolve_apply",
     "evolve_apply_conflicted_pr", "evolve_apply_curator_report",